Create Master_Shiny_Example
This Shiny example creates a histogram with mean = 0 and standard deviation = 1. Shiny allows the user to chose the number of samples, number of breaks, there is a check box if the user wants to see the values, there are two interactive text boxes (one for the title of the plot, one for the x-axis), and a set of three checkboxes to determine the color of the plot. This example gives a small taste of what's possible with Shiny :)

PS - this was a lot of fun to create!