Create Isolate_example
This is an interactive histogram. The user can choose the number of samples from 1 to 10,000, the number of breaks from 1 to 1,000, add a title, and Shiny provides a statistical summary of the data in the distribution. Finally Shiny also draws the histogram.

One point worth noting is that changing the number of breaks does not change the summary statistics, but changing the number of items does change the summary statistics for the histogram.

Updating or changing any of the values instantly updates the histogram.