Create Master_template
I was very surprised that I never saw this in all my Shiny training. It's a master template that includes everything that Shiny can do in one file. Obviously no one will use all of this at the same time, but it makes it all available, eliminates any spelling errors, and reminds me of the elements of a Shiny file.

It's super easy to edit this to make a live Shiny file, and can also point out cool things I might not have considered.

This was fun to do, and I use it as I make other Shiny files, since it has pretty much everything Shiny can do!